[FIX] Revise Code of Conduct
We have, as a group, revised the Code of Conduct for the A7 submission as per the feedback in issue #4.

Additions to 'How do we ensure that everyone in our group stays informed about the individual contributions?' and 'How do we ensure knowledge transfer among our team members?':

To ensure that each team member is informed on the technical/code level we will use workshops, git commit messages and Trello. The workshops will be used to transfer knowledge between team members and help team members stay informed on the technical level. To ensure that every team member stays informed about the individual contributions the git commit messages will be connected to issues that have been assigned to specific team members, and the team members that contributed to the commit will be credited as per the defined commit message format (see README.md). Team members will also use Trello to stay up to date with the progress of tasks.

Additions to 'How will we solve conflicts?':

Conflict resolution has been improved by adding a list of steps for how to resolve the conflict, under 'Discussion between group members', since the original version did not include any details for how the team will solve conflicts. The list of confluct resolution steps was found at https://amanet.org/articles/the-five-steps-to-conflict-resolution. These steps will serve as guidelines if a conflict arises.

Resolves: #4
Co-authored-by: chenr <chenr@student.chalmers.se>
Co-authored-by: shonaigh <shonaigh@student.chalmers.se>
Co-authored-by: hashmati <hashmati@student.chalmers.se>