[FIX] Update commit message example
Co-authored-by: chenr <chenr@student.chalmers.se>
Co-authored-by: shonaigh <shonaigh@student.chalmers.se>
Co-authored-by: hashmati <hashmati@student.chalmers.se>

Updated the commit message example in the README to reflect the format we
have agreed to use. Moved the co-authored-by tag to the top of the commit
message and changed the format to 'chalmersID' <chalmersID@student.chalmers.se>.
Also added a [TEST] tag for tests and CI/CD related commits.

Resolves: #4