[IMPLEMENT] HSV color detection
Co-authored-by: nordje <nordje@student.chalmers.se>
Co-authored-by: shonaigh <shonaigh@student.chalmers.se>

Tomato detection example integrated with our application and the provided
.rec files. The code is opening up the sliders for the HSV ranges and the
color-space image. We did this to find the HSV color ranges for the blue
and yellow cones. This was done by creating trackbars (the sliders) and
by creating a Mat object for the color-space image and the HSV values.
It also converts the image from the video to an HSV image.

Resolves #9