[IMPLEMENT] Add boundary boxes to detected cones
Co-authored-by: hashmati <hashmati@student.chalmers.se>
Co-authored-by: nordje <nordje@student.chalmers.se>

Used the colour mask to enable contour finding and put boundary boxes on top
of found contours. This will later allow us to provide the locations of the
cones to the steering algorithm. A window will open with trackbar sliders
allowing the user to enter HSV values of the cones to be tracked. This will
be highlighed in another window with coloured rectangles and circles.

Resolves: #9