[IMPLEMENT] Filter for both cone colors
Co-authored-by: hashmati <hashmati@student.chalmers.se>
Co-authored-by: nordje <nordje@student.chalmers.se>

Improved the color detection for cones. Removed the trackbar window,
the randomly generated colors and the inner circles of the bounding
boxes. The trackbars were just for finding HSV ranges for the two colors
and could therefore be removed. The randomly generated colors were too
distracting and are replaced with single colored bounding boxes. The inner
circles of the bounding boxes were removed too because we just need the
bounding boxes.

Added both blue and yellow HSV detection for the same image/window, this
means that the system shows both blue and yellow cones with different
colored bounding boxes. This is so we are able to separate blue and yellow
cones from each other which might be helpful for the steering wheel
algorithm later on.

Resolves: #9