[FIX] Update README set-up instructions and workflow
Updated the README set-up section to include docker build and run
instructions. This is because we usually build and run the system using
docker and not manually.

Also updated the workflow section to include clarify when unit tests are
supposed to be written and when a branch can be merged to master, i.e.,
when the CI passes and the feature has passing unit tests.

Resolves: #6, #11