[IMPLEMENT] Add region of interest
Co-authored-by: chenr <chenr@student.chalmers.se>
Co-authored-by: shonaigh <shonaigh@student.chalmers.se>
Co-authored-by: nordje <nordje@student.chalmers.se>

We added the region of interst to the video feed to help filter out some
of the noise. The noise was filtered out to improve the tracking of the
necessary cones and speed up the application. This was done by cropping
the image using coordinates and changing the width and height of the image.

Resolves #12