[FIX] Update README to include feedback
Updated the set-up docker instructions to include how to add data to the 
shared memory so that the application won't crash, i.e., opening opendlv-
vehicle-view in Chrome and playing a selected recording before starting 
our application.

Added a point (4.1) to the workflow to clarify who's responsible for writing 
the unit tests for a new feature.

Resolves: #6, #11.