[IMPLEMENT] Add noise filtering
Co-authored-by: chenr <chenr@student.chalmers.se>
Co-authored-by: shonaigh <shonaigh@student.chalmers.se>
Co-authored-by: nordje <nordje@student.chalmers.se>

We added the code for filtering out the noise around the cones. This was
done by creating a new method called filtering, that handles the filtering
with eroding and dilating objects and holes in the foreground.
We are filtering out the noise to improve the detection of the cones
and limit the amount of false positives found.

Resolves #13