[IMPLEMENT] Verbose mode
Implemented the --verbose mode that enables the application to open a
debug window when --verbose is added to the run command. This was done
so that it fulfills the requirement for the application and so that
relevant information about the application can be displayed. This was
done by first creating a new Mat object, then using putText to
display text on the Mat object for the debug window. The window is
then shown by using imShow.

Resolves: #20