[IMPLEMENT] Print GSR with timestamp to console
Co-authored-by: shonaigh <shonaigh@student.chalmers.se>

We decided to use an alternative method to extract the data points to CSV
files instead of automating it using the Data Recorder. We will now extract
the console output using a command line command which outputs a txt file
with the console data. We can then open the textfile and remove everything
that is not data before converting the file into a CSV file which can be
used to plot graphs etc.

The command line output had been adjusted to output what is required for
A20 and A24. You can modify this if you are interested in other points of
data.

We commented out the previous ground steering request console outputs but
they should be removed at a later date.

Resolves: #19