[FIX] Stop crashing when not using verbose mode
Co-authored-by: shonaigh <shonaigh@student.chalmers.se>

Removed any windows opened outside of the VERBOSE if statement because
they would cause the application to crash whenever -e DISPLAY=:0
was not provided in the docker commands.

Removed the debug window since the requirement was for debug information
and not necessarily the need to open a separate window with data. We will
instead display necessary debug data (such as bounding boxes and
coordinates) on the video window itself.

Resolves: #20