[FIX] Steering Algorithm Calculations
Co-authored-by: chenr <chenr@student.chalmers.se>

Removed the steeringWheelDirection method since that responsibility was
moved to the steeringWheelAngle method. Fixed the steeringWheelAngle
method by dividing it into two sections, one for the left side of the ROI
and one for the right side of the ROI. This is because the algorithm was
producing wrong positive and negative numbers for cones and sometimes a
steering angle which was larger than the maximum value.

Also added in comments in the DataProcessor.cpp file for both methods.

Updated the small ROI values.

Resolves: #23, #25